[PHP] Solve isogram using simple loop and set
I see some other solutions using array_unique in combination with count.
But that would require walking the input string twice, once for unique
and once to measure the length, in addition to walking the unique array
once to measure it's length.
Whereas my code only walks the input once, and stops at the first duplicate.